This is the first push I've made after a leave of abscence from this project. All previous pushes were made from archived version of this project at landmark stages of development. Pushes after this one should generally be smaller and more incremental. This push is also sweeping however.
-Refactored the TypeManager class used to retrieve the expected json signatures of our class types. Attributes no longer need to be redeclared in the signatures of child classes. This system could be made more efficient and there are comments about a more effiecient implementation, but that has presently been eschewed because it causes false positives in CRTDBG's memory leak detection.
-Actions now are classified as continuous or not continuous. Actions that are not continuous are removed from the action queue after the EventQueue finishes updating in an iteration of the game loop.
-Fixed a bug in SLists where the move assignment operator was always setting size to 0.
-Began organizing files into VS filters.

+++ source/Library.Shared/DefaultHashFunctor.cpp
@@ -9,7 +9,7 @@ namespace FIEAGameEngine
 		size_t hashValue = 0;
 		for (size_t i = 0; i < size; i++)
 		{
-			hashValue += data[i] * 33;
+			hashValue += data[i] * size_t(33);
 		}
 
 		return hashValue;
